* Panasonic DC-GX7MK3 camera support
* SONY ILCE-7C camera support

* upstream/develop:
  Github Actions CI: enable MINGW32 testing
  iPoint2D: `size_t` is not guaranteed to be larger than `int`
  GitHub Actions CI: update to LLVM/Clang 11
  Cr2Decoder::decodeNewFormat(): actually ensure that subsampled image dims are multiples of subsampling factor
  Github Actions CI: Add CodeQL analysis
  Port clang static analysis ci job to github actions/CodeChecker
  Cr2Decoder::decodeNewFormat(): ensure that if subsampling is on, there is subsampling, and vice versa
  AbstractLJpegDecompressor::parseSOF(): don't write mRaw->metadata.subsampling, but check against it
  Cr2Decoder::decodeNewFormat(): store subsampling in RawImage's subsampling field
  rawspeed: SONY ILCE-7C support
  Add DC-GX7MK3 alias & remove spaces
  GHA: try to actually fix rpu-enabled coverage build
  And fix yet another typo in gha yaml
  GH Actions: fix rpu-less build
  Github actions: and don't forget to re-enable linux clang build...
  Github actions: and don't forget to re-enable linux RelWithDebInfo build..
  Fixup github actions workflow yaml
  Port modern gcc/clang/gcc coverage builds to github actions
  README: update for travis-ci org->com migration
  Github Actions CI: retry using better upload name for codecov uploads

+++ src/librawspeed/common/Point.h
@@ -22,15 +22,15 @@
 #pragma once
 
 #include <algorithm>   // for max, min
-#include <cstddef>     // for size_t
+#include <cstdint>     // for int32_t, uint64_t
 #include <type_traits> // for make_signed, make_signed<>::type
 
 namespace rawspeed {
 
 class iPoint2D {
 public:
-  using value_type = int;
-  using area_type = size_t;
+  using value_type = int32_t;
+  using area_type = uint64_t;
 
   constexpr iPoint2D() = default;
   constexpr iPoint2D(value_type a, value_type b) : x(a), y(b) {}
